// src/api/generated/models/adminReviewDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Bookora API
 * Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.
 * OpenAPI spec version: 1.0
 */
import type { ReviewProductDto } from './reviewProductDto';

export interface AdminReviewDto {
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  id: string;
  orderId: string;
  orderCode: string;
  product: ReviewProductDto;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  content?: string | null;
  isVisible: boolean;
  createdAt: string;
  updatedAt: string;
  customerName: string;
  customerEmail: string;
  /** @nullable */
  customerAvatarUrl?: string | null;
  /**
     * @minLength 26
     * @maxLength 26
     * @pattern ^[0-7][0-9A-HJKMNP-TV-Z]{25}$
     */
  branchId: string;
  branchName: string;
}